chromium/third_party/blink/web_tests/wpt_internal/display-lock/css-subtree-visibility/inheritance.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Subtree Visibility: subtree-visibility inheritance</title>
<link rel="author" title="Vladimir Levin" href="mailto:vmpstr@chromium.org">
<meta name="assert" content="subtree-visibility does not inherit">
<meta name="assert" content="subtree-visibility has initial value 'visible'">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
<div id="container">
<div id="target"></div>
</div>
<script>
assert_not_inherited("subtree-visibility", "visible", "auto");
</script>
</body>
</html>
